requirements.txt
python-dotenv
langchain
langchain-community
langchain-groq
PyPDF2
faiss-cpu
streamlit
sentence-transformers
pandas
plotly
cohere==4.57
-e .